Include changes in client configuration

+++ client/src/To-Do-List.js
@@ -71,9 +71,9 @@ class ToDoList extends Component {
         }
       )
       .then((res) => {
+        this.getTask();
         console.log(res);
       });
-      this.getTask();
   };
 
   undoTask = (id) => {
@@ -88,36 +88,41 @@ class ToDoList extends Component {
         }
       )
       .then((res) => {
+        this.getTask();
         console.log(res);
       });
-      this.getTask();
   };
 
   deleteTask = (id) => {
-    axios
-      .delete(endpoint + "/api/deleteTask/" + id, {
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-      })
-      .then((res) => {
-        console.log(res);
-      });
-      this.getTask();
+    if (window.confirm("Are you sure you want to delete this task?")) {
+      axios
+        .delete(endpoint + "/api/deleteTask/" + id, {
+          headers: {
+            "Content-Type": "application/x-www-form-urlencoded",
+          },
+        })
+        .then((res) => {
+          console.log(res);
+          this.getTask();
+        })
+        .catch((err) => console.log(err));
+    }
   };
 
   deleteAllTasks = () => {
-    axios
-      .delete(endpoint + "/api/deleteAllTasks", {
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-      })
-      .then((res) => {
-        console.log(res);
-        this.getTask();
-      })
-      .catch((err) => console.log(err));
+    if (window.confirm("Are you sure you want to delete all tasks?")) {
+      axios
+        .delete(endpoint + "/api/deleteAllTasks", {
+          headers: {
+            "Content-Type": "application/x-www-form-urlencoded",
+          },
+        })
+        .then((res) => {
+          console.log(res);
+          this.getTask();
+        })
+        .catch((err) => console.log(err));
+    }
   };
   
 
@@ -164,11 +169,12 @@ class ToDoList extends Component {
   };
 
   render() {
+    const taskCount = this.state.items.length;
     return (
       <div>
         <div className="row">
           <Header className="header" as="h2" color="black">
-            TO DO LIST
+            TO DO APP
           </Header>
         </div>
         <div className="row">
@@ -189,381 +195,27 @@ class ToDoList extends Component {
         </div>
         <div className="row">
           <Card.Group>{this.renderItems()}</Card.Group>
-        </div>
-        <Button
+        </div >
+        <div className="row form-container">
+          <p className="task-counter">You have {taskCount} pending tasks</p>
+          <Button
             color="purple"
-            style={{ marginLeft: "10px" }}
+            style={{ marginLeft: "auto" }}
             onClick={this.deleteAllTasks}
           >
-            Clear All
+            Delete All
           </Button>
+        </div>
+        {/* <Button
+            color="purple"
+            style={{ marginLeft: "949px" }}
+            onClick={this.deleteAllTasks}
+          >
+            Delete All
+          </Button> */}
       </div>
     );
   }
 }
 
 export default ToDoList;
-
-
-// import React, { Component } from "react";
-// import axios from "axios"; // It makes the API calls to the backend
-// import { Card, Header, Form, Input, Icon, Button } from "semantic-ui-react"; // Import Button
-// let endpoint = "http://localhost:9000";
-
-// class ToDoList extends Component {
-//   constructor(props) {
-//     super(props);
-
-//     this.state = {
-//       task: "",
-//       items: [],
-//     };
-//   }
-
-//   componentDidMount() {
-//     this.getTask();
-//   }
-
-//   onChange = (event) => {
-//     this.setState({
-//       [event.target.name]: event.target.value,
-//     });
-//   };
-
-//   onSubmit = (event) => {
-//     event.preventDefault(); // Prevent default form submission behavior
-//     let { task } = this.state;
-
-//     if (task) {
-//       axios
-//         .post(
-//           endpoint + "/api/task",
-//           { task: task }, // Ensure the task is correctly sent
-//           {
-//             headers: {
-//               "Content-Type": "application/x-www-form-urlencoded",
-//             },
-//           }
-//         )
-//         .then((res) => {
-//           this.getTask();
-//           this.setState({
-//             task: "",
-//           });
-//           console.log(res);
-//         });
-//     }
-//   };
-
-//   getTask = () => {
-//     axios.get(endpoint + "/api/task").then((res) => {
-//       if (res.data) {
-//         this.setState({ items: res.data });
-//       } else {
-//         this.setState({ items: [] });
-//       }
-//     });
-//   };
-
-//   updateTask = (id) => {
-//     axios
-//       .put(
-//         endpoint + "/api/task/" + id,
-//         {},
-//         {
-//           headers: {
-//             "Content-Type": "application/x-www-form-urlencoded",
-//           },
-//         }
-//       )
-//       .then((res) => {
-//         console.log(res);
-//       });
-//       this.getTask();
-//   };
-
-//   undoTask = (id) => {
-//     axios
-//       .put(
-//         endpoint + "/api/undoTask/" + id,
-//         {},
-//         {
-//           headers: {
-//             "Content-Type": "application/x-www-form-urlencoded",
-//           },
-//         }
-//       )
-//       .then((res) => {
-//         console.log(res);
-//       });
-//       this.getTask();
-//   };
-
-//   deleteTask = (id) => {
-//     axios
-//       .delete(endpoint + "/api/deleteTask/" + id, {
-//         headers: {
-//           "Content-Type": "application/x-www-form-urlencoded",
-//         },
-//       })
-//       .then((res) => {
-//         console.log(res);
-//       });
-//       this.getTask();
-//   };
-
-//   renderItems = () => {
-//     return this.state.items.map((item) => {
-//       let color = "yellow";
-//       let style = {
-//         wordWrap: "break-word",
-//       };
-//       if (item.status) {
-//         color = "green";
-//         style["textDecorationLine"] = "line-through";
-//       }
-//       return (
-//         <Card key={item._id} color={color} fluid className="rough">
-//           <Card.Content style={{ backgroundColor: "#ebebeb" }}>
-//             <Card.Header textAlign="left">
-//               <div style={style}>{item.task}</div>
-//             </Card.Header>
-//             <Card.Meta textAlign="right">
-//               <Icon
-//                 name="check circle"
-//                 color="green"
-//                 onClick={() => this.updateTask(item._id)}
-//               />
-//               <span style={{ paddingRight: 10 }}>Done</span>
-//               <Icon
-//                 name="undo"
-//                 color="orange"
-//                 onClick={() => this.undoTask(item._id)}
-//               />
-//               <span style={{ paddingRight: 10 }}>Undo</span>
-//               <Icon
-//                 name="delete"
-//                 color="red"
-//                 onClick={() => this.deleteTask(item._id)}
-//               />
-//               <span style={{ paddingRight: 10 }}>Delete</span>
-//             </Card.Meta>
-//           </Card.Content>
-//         </Card>
-//       );
-//     });
-//   };
-
-//   render() {
-//     return (
-//       <div>
-//         <div className="row">
-//           <Header className="header" as="h2" color="black">
-//             TO DO LIST
-//           </Header>
-//         </div>
-//         <div className="row">
-//           <Form onSubmit={this.onSubmit}>
-//             <Input
-//               type="text"
-//               name="task"
-//               onChange={this.onChange}
-//               value={this.state.task}
-//               fluid
-//               placeholder="Add your new task"
-//             />
-//             <Button type="submit" color="blue" style={{ marginTop: "10px" }}>
-//               Add Task
-//             </Button>
-//           </Form>
-//         </div>
-//         <div className="row">
-//           <Card.Group>{this.renderItems()}</Card.Group>
-//         </div>
-//       </div>
-//     );
-//   }
-// }
-
-// export default ToDoList;
-
-//--------
-
-
-// import React, { Component } from "react";
-// import axios from "axios"; // It makes the API calls to the backend
-// import { Card, Header, Form, Input, Icon } from "semantic-ui-react";
-// let endpoint = "http://localhost:9000";
-
-// class ToDoList extends Component {
-//   constructor(props) {
-//     super(props);
-
-//     this.state = {
-//       task: "",
-//       items: [],
-//     };
-//   }
-
-//   componentDidMount() {
-//     this.getTask();
-//   }
-
-//   onChange = (event) => {
-//     this.setState({
-//       [event.target.name]: event.target.value,
-//     });
-//   };
-
-//   onSubmit = () => {
-//     let { task } = this.state;
-
-//     if (task) {
-//       axios
-//         .post(
-//           endpoint + "/api/task",
-//           { task: task }, // Ensure the task is correctly sent
-//           {
-//             headers: {
-//               "Content-Type": "application/x-www-form-urlencoded",
-//             },
-//           }
-//         )
-//         .then((res) => {
-//           this.getTask();
-//           this.setState({
-//             task: "",
-//           });
-//           console.log(res);
-//         });
-//     }
-//   };
-
-//   getTask = () => {
-//     axios.get(endpoint + "/api/task").then((res) => {
-//       if (res.data) {
-//         this.setState({ items: res.data });
-//       } else {
-//         this.setState({ items: [] });
-//       }
-//     });
-//   };
-
-// updateTask = (id) => {
-//     axios
-//       .put(
-//         endpoint + "/api/task/" + id,
-//         {},
-//         {
-//           headers: {
-//             "Content-Type": "application/x-www-form-urlencoded",
-//           },
-//         }
-//       )
-//       .then((res) => {
-//         console.log(res);
-//       })
-//       this.getTask();
-//   };
-
-//   undoTask = (id) => {
-//     axios
-//       .put(
-//         endpoint + "/api/undoTask/" + id,
-//         {},
-//         {
-//           headers: {
-//             "Content-Type": "application/x-www-form-urlencoded",
-//           },
-//         }
-//       )
-//       .then((res) => {
-//         console.log(res);
-//       })
-//       this.getTask();
-//   };
-
-//   deleteTask = (id) => {
-//     axios
-//       .delete(endpoint + "/api/deleteTask/" + id, {
-//         headers: {
-//           "Content-Type": "application/x-www-form-urlencoded",
-//         },
-//       })
-//       .then((res) => {
-//         console.log(res);
-//       })
-//       this.getTask();
-//   };
-
-//   renderItems = () => {
-//     return this.state.items.map((item) => {
-//       let color = "yellow";
-//       let style = {
-//         wordWrap: "break-word",
-//       };
-//       if (item.status) {
-//         color = "green";
-//         style["textDecorationLine"] = "line-through";
-//       }
-//       return (
-//         <Card key={item._id} color={color} fluid className="rough">
-//           <Card.Content style={{ backgroundColor: "#ebebeb" }}>
-//             <Card.Header textAlign="left">
-//               <div style={style}>{item.task}</div>
-//             </Card.Header>
-//             <Card.Meta textAlign="right">
-//               <Icon
-//                 name="check circle"
-//                 color="green"
-//                 onClick={() => this.updateTask(item._id)}
-//               />
-//               <span style={{ paddingRight: 10 }}>Done</span>
-//               <Icon
-//                 name="undo"
-//                 color="orange"
-//                 onClick={() => this.undoTask(item._id)}
-//               />
-//               <span style={{ paddingRight: 10 }}>Undo</span>
-//               <Icon
-//                 name="delete"
-//                 color="red"
-//                 onClick={() => this.deleteTask(item._id)}
-//               />
-//               <span style={{ paddingRight: 10 }}>Delete</span>
-//             </Card.Meta>
-//           </Card.Content>
-//         </Card>
-//       );
-//     });
-//   };
-
-//   render() {
-//     return (
-//       <div>
-//         <div className="row">
-//           <Header className="header" as="h2" color="black" >
-//             TO DO LIST
-//           </Header>
-//         </div>
-//         <div className="row">
-//           <Form onSubmit={this.onSubmit}>
-//             <Input
-//               type="text"
-//               name="task"
-//               onChange={this.onChange}
-//               value={this.state.task}
-//               fluid
-//               placeholder="Add your new task"
-//             />
-//           </Form>
-//         </div>
-//         <div className="row">
-//           <Card.Group>{this.renderItems()}</Card.Group>
-//         </div>
-//       </div>
-//     );
-//   }
-// }
-
-// export default ToDoList;
